# docs/api/openapi-f02.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access V1 y usuario; refresh V1 en cookie HttpOnly refresh_token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "accessToken",
                    "user"
                  ],
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "Bearer V1; refresh se entrega solo en cookie HttpOnly."
                    },
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "email",
                        "role",
                        "tenantId",
                        "tenant"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        },
                        "tenant": {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DTO invalido."
          },
          "401": {
            "description": "Credenciales o contexto invalidos."
          },
          "429": {
            "description": "Limite de intentos excedido."
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Access V1 renovado; refresh V1 rotado en cookie HttpOnly.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "accessToken",
                    "user"
                  ],
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "Bearer V1; refresh se entrega solo en cookie HttpOnly."
                    },
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "email",
                        "role",
                        "tenantId",
                        "tenant"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        },
                        "tenant": {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh V1 invalido o revocado."
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Renovar sesión con refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Refresh V1 revocado; cookie eliminada."
          },
          "401": {
            "description": "Access invalido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar sesión (revoca refresh token del dispositivo)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil revalidado con Member actual.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "tenantId",
                    "tenant"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "tenantId": {
                      "type": "string"
                    },
                    "tenant": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "slug"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Contexto inactivo o invalido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener perfil del usuario autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/v2/login": {
      "post": {
        "description": "Contrato versionado 2. COOKIE es el transporte Web por defecto; BODY prepara clientes Mobile futuros.",
        "operationId": "AuthV2Controller_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginV2Dto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión V2 creada.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "contractVersion",
                    "accessToken",
                    "refreshTransport",
                    "sessionExpiresAt",
                    "context"
                  ],
                  "properties": {
                    "contractVersion": {
                      "type": "integer",
                      "enum": [
                        2
                      ]
                    },
                    "accessToken": {
                      "type": "string",
                      "description": "Bearer JWT con memberId, sessionId y ver=2."
                    },
                    "refreshTransport": {
                      "type": "string",
                      "enum": [
                        "COOKIE",
                        "BODY"
                      ]
                    },
                    "sessionExpiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "context": {
                      "type": "object",
                      "required": [
                        "account",
                        "member",
                        "tenant",
                        "role",
                        "session"
                      ],
                      "properties": {
                        "account": {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "email",
                            "isActive"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "isActive": {
                              "type": "boolean"
                            }
                          }
                        },
                        "member": {
                          "type": "object",
                          "required": [
                            "id",
                            "role",
                            "isActive"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "AGRICULTOR",
                                "DIRECTIVA",
                                "ADMINISTRADOR"
                              ]
                            },
                            "isActive": {
                              "type": "boolean"
                            }
                          }
                        },
                        "tenant": {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "slug",
                            "isActive"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            },
                            "isActive": {
                              "type": "boolean"
                            }
                          }
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "AGRICULTOR",
                            "DIRECTIVA",
                            "ADMINISTRADOR"
                          ]
                        },
                        "session": {
                          "type": "object",
                          "required": [
                            "id",
                            "status",
                            "refreshTransport",
                            "createdAt",
                            "lastSeenAt",
                            "expiresAt",
                            "clientRegistrationId"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "ACTIVE"
                              ],
                              "description": "El contexto se entrega solo para sesiones activas; el modelo persistido tambien contempla REVOKED y COMPROMISED."
                            },
                            "refreshTransport": {
                              "type": "string",
                              "enum": [
                                "COOKIE",
                                "BODY"
                              ]
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "lastSeenAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "expiresAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "clientRegistrationId": {
                              "type": "string",
                              "nullable": true
                            }
                          }
                        }
                      }
                    },
                    "refreshToken": {
                      "type": "string",
                      "description": "Solo aparece con transporte BODY; COOKIE usa refresh_token_v2 HttpOnly."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "DTO o transporte invalido."
          },
          "401": {
            "description": "Credenciales o contexto inactivos."
          },
          "429": {
            "description": "Limite de intentos excedido."
          }
        },
        "summary": "Iniciar una sesión auth V2",
        "tags": [
          "auth v2"
        ]
      }
    },
...
    "/api/v1/auth/v2/logout": {
      "post": {
        "description": "Idempotente incluso sin credencial y utilizable aunque el access haya expirado. Cookie y BODY son alternativas; OpenAPI usa {} para BODY o ausencia de credencial. No revoca ClientRegistration.",
        "operationId": "AuthV2Controller_logout",
        "parameters": [],
        "requestBody": {
          "required": false,
          "description": "Opcional: refresh_token_v2 por cookie o refreshToken por BODY. Sin credencial, logout es idempotente (204). Ambas fuentes causan 400.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutV2Dto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session V2 revocada; cookie eliminada."
          },
          "400": {
            "description": "Transporte ambiguo o DTO invalido."
          }
        },
        "security": [
          {
            "refresh_token_v2": []
          },
          {}
        ],
        "summary": "Revocar la sesión V2 actual",
        "tags": [
          "auth v2"
        ]
      }
    },
    "/api/v1/auth/v2/me": {
      "get": {
        "description": "Revalida Account, Member, Tenant, rol, Session y ClientRegistration asociado.",
        "operationId": "AuthV2Controller_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Contexto V2 actual revalidado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "account",
                    "member",
                    "tenant",
                    "role",
                    "session"
                  ],
                  "properties": {
                    "account": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "email",
                        "isActive"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "isActive": {
                          "type": "boolean"
                        }
                      }
                    },
                    "member": {
                      "type": "object",
                      "required": [
                        "id",
                        "role",
                        "isActive"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "AGRICULTOR",
                            "DIRECTIVA",
                            "ADMINISTRADOR"
                          ]
                        },
                        "isActive": {
                          "type": "boolean"
                        }
                      }
                    },
                    "tenant": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "slug",
                        "isActive"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "isActive": {
                          "type": "boolean"
                        }
                      }
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "AGRICULTOR",
                        "DIRECTIVA",
                        "ADMINISTRADOR"
                      ]
                    },
                    "session": {
                      "type": "object",
                      "required": [
                        "id",
                        "status",
                        "refreshTransport",
                        "createdAt",
                        "lastSeenAt",
                        "expiresAt",
                        "clientRegistrationId"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "ACTIVE"
                          ],
                          "description": "El contexto se entrega solo para sesiones activas; el modelo persistido tambien contempla REVOKED y COMPROMISED."
                        },
                        "refreshTransport": {
                          "type": "string",
                          "enum": [
                            "COOKIE",
                            "BODY"
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "lastSeenAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "clientRegistrationId": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Contexto actual inactivo o revocado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener contexto vigente de la sesión V2",
        "tags": [
          "auth v2"
        ]
      }
    },
    "/api/v1/auth/v2/session/client": {
      "post": {
        "description": "Deriva session/member/tenant del access V2. El mismo binding es idempotente y un rebind distinto responde 409.",
        "operationId": "AuthV2Controller_bindClient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindSessionClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Binding idempotente a ClientRegistration propio activo.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "registrationId",
                    "status"
                  ],
                  "properties": {
                    "registrationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "ACTIVE"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Registration ID invalido."
          },
          "401": {
            "description": "Session o ClientRegistration revocados."
          },
          "404": {
            "description": "Registro no visible."
          },
          "409": {
            "description": "La sesión ya está ligada a otro registro."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Asociar la sesión V2 a un ClientRegistration activo",
        "tags": [
          "auth v2"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agrochemical Inventory API",
    "description": "API central de identidad, membresia, contratos autorizados, campanas y catalogos compartidos, con autenticacion versionada. Algunos endpoints legacy permanecen por compatibilidad; su presencia en OpenAPI no los convierte en arquitectura target del Proyecto de Grado.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refresh_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      },
      "refresh_token_v2": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token_v2"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@agro.local"
          },
          "password": {
            "type": "string",
            "example": "Admin123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginV2Dto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@agro.local"
          },
          "password": {
            "type": "string",
            "example": "Admin123!"
          },
          "refreshTransport": {
            "type": "string",
            "enum": [
              "COOKIE",
              "BODY"
            ],
            "default": "COOKIE",
            "description": "COOKIE para Web; BODY prepara clientes Mobile futuros."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "LogoutV2Dto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh opaco V2. Exclusivo del transporte BODY."
          }
        }
      },
      "BindSessionClientDto": {
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid",
            "description": "ID central de ClientRegistration."
          }
        },
        "required": [
          "registrationId"
        ]
      },
...
    }
  }
}